moved new map functions from map.js to mapFunctions.js
