+++ front-end/src/Components/Map/map.js
@@ -2,7 +2,7 @@ import React from 'react';
 import ReactNYC from 'react-nyc-choropleth';
 import { API, graphqlOperation } from "aws-amplify";
 import * as queries from '../../graphql/queries';
-import { getColor } from './mapFunctions';
+import { getColor, averageTipsByNeighborhood, averageTipsClean } from './mapFunctions';
 
 class Map extends React.Component {
   constructor(props) {
@@ -29,40 +29,7 @@ class Map extends React.Component {
     const mapboxType = "light";
     const position = [40.7831, -73.9712];
     const zoom = 12;
-    
-    const averageTipsByNeighborhood = (tip_info) => {
-      const neighborhoods = new Object();
-      tip_info.forEach(entry => {
-          let hoodName = entry.neighborhood;
-          if (! neighborhoods.hasOwnProperty(hoodName)) {
-              neighborhoods[hoodName] = {neighborhood: hoodName,
-                                      tipsPerHour: (entry.takehome / entry.shift_length),
-                                      totalHour: entry.shift_length};
-          }
-          else {
-              let hours = (neighborhoods[hoodName].totalHour + entry.shift_length);
-              neighborhoods[hoodName].tipsPerHour = neighborhoods[hoodName].tipsPerHour * neighborhoods[hoodName].totalHour / hours + entry.takehome / hours;
-              neighborhoods[hoodName].totalHour = hours;
-          }
-      });
-      return neighborhoods;
-    }
 
-    const averageTipsClean = (average_tips) => {
-      let data = [];
-      let element = {};
-      for (var entry in average_tips) {
-        element = {};
-        element.name = average_tips[entry].neighborhood;
-        element.values = [];
-        element.values.push({label: "Avg Hourly $", val: Math.floor(average_tips[entry].tipsPerHour)});
-        console.log(element.values[0].val);
-        element.color = getColor(element.values[0].val);
-        data.push(element);
-      }
-      return data;
-    }
-    
     const neighborhoodStyle = { weight: 1, opacity: 1, color: '#666', dashArray: '3', fillOpacity: 0.7 };
     const neighborhoodHoverStyle = { weight: 5, color: '#FFF', dashArray: '1', fillOpacity: 0.7 };
     const excludeNeighborhoods = ["Liberty Island", "Ellis Island"];
@@ -70,12 +37,11 @@ class Map extends React.Component {
     const data = averageTipsClean(averageTipsByNeighborhood(this.props.tip_info));
     //const data = averageTipsClean(averageTipsByNeighborhood(this.state.tip_data))
 
-
     return (
       <div>
         <ReactNYC
           mapboxAccessToken={mapboxAccessToken} // Required
-          mapHeight="800px" // Required
+          mapHeight="600px" // Required
           mapWidth="100%"
           className="container"
           mapboxType={mapboxType}
